Переносит общую логику обновления
