+++ src/main.js
@@ -17,5 +17,4 @@ render(mainElement, filterComponent, renderPosition.AFTERBEGIN);
 render(mainElement, infoComponent, renderPosition.AFTERBEGIN);
 
 const tripPresenter = new TripPresenter(points);
-debugger
 tripPresenter.init(points);
